feat(spaceLayout): set SpaceTopHeader to whole /spaces/... page
