+++ app/spaces/[spaceid]/edit/layout.tsx
@@ -6,28 +6,12 @@ import { CaretUpDown } from '@phosphor-icons/react';
 import PcSidebar from './components/settingSidebar/PcSidebar';
 import MobileSidebar from './components/settingSidebar/MobileSidebar';
 import BackHeader from './components/backHeader';
-import SpaceEditHeader from './components/SpaceEditHeader';
+import SpaceTopHeader from '../components/SpaceTopHeader';
 import { useQuery } from '@tanstack/react-query';
 import { useCeramicContext } from '@/context/CeramicContext';
 import { getSpacesQuery } from '@/services/space';
 import { Space } from '@/types';
 
-// Space query by ID
-const getSpaceByIdQuery = (id: string) => `
-  query GetSpace($id: ID!) {
-    node(id: $id) {
-      ...on ZucitySpace {
-        id
-        avatar
-        name
-        description
-        banner
-      }
-    }
-  }
-`;
-
-// 标题映射表
 const TITLE_MAP: Record<string, string> = {
   overview: 'Community Overview',
   access: 'Access Management',
@@ -43,35 +27,6 @@ const SpaceEditLayout = ({ children }: { children: React.ReactNode }) => {
   const params = useParams();
   const spaceId = params.spaceid.toString();
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-  const [isClient, setIsClient] = useState(false);
-  const { composeClient } = useCeramicContext();
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  // Fetch space data
-  const { data: spaceData } = useQuery({
-    queryKey: ['space', spaceId],
-    queryFn: async () => {
-      try {
-        const response = await composeClient.executeQuery(
-          getSpaceByIdQuery(spaceId),
-          {
-            id: spaceId,
-          },
-        );
-        if (response.data && response.data.node) {
-          return response.data.node as Space;
-        }
-        return null;
-      } catch (error) {
-        console.error('Failed to fetch space:', error);
-        return null;
-      }
-    },
-    enabled: !!spaceId && !!composeClient,
-  });
 
   const getCurrentTitle = () => {
     if (pathname === `/spaces/${spaceId}/edit`) {
@@ -85,56 +40,42 @@ const SpaceEditLayout = ({ children }: { children: React.ReactNode }) => {
   };
 
   return (
-    <>
-      {/* Mobile Header - Replaces global header on mobile */}
-      {isClient && (
-        <div className="pc:hidden tablet:hidden mobile:block">
-          <SpaceEditHeader
-            spaceId={spaceId}
-            spaceName={spaceData?.name || 'Community'}
-            spaceAvatar={spaceData?.avatar}
-            title={getCurrentTitle()}
-          />
+    <div className="flex h-[calc(100vh-50px)]">
+      <PcSidebar currentPath={pathname} hasChanges={false} />
+
+      <MobileSidebar
+        isOpen={isMobileMenuOpen}
+        onClose={() => setIsMobileMenuOpen(false)}
+        currentPath={pathname}
+      />
+
+      <div className="flex-1 flex flex-col">
+        {/* PC */}
+        <div className="pc:flex tablet:hidden mobile:hidden h-[50px] border-[#2C2C2C] border-b border-[rgba(255,255,255,0.1)] flex items-center px-5 backdrop-blur-[40px] ">
+          <h1 className="text-[18px] font-bold text-white shadow-[0px_5px_10px_0px_rgba(0,0,0,0.15)]">
+            {getCurrentTitle()}
+          </h1>
         </div>
-      )}
 
-      <div className="flex h-[calc(100vh-50px)]">
-        <PcSidebar currentPath={pathname} hasChanges={false} />
-
-        <MobileSidebar
-          isOpen={isMobileMenuOpen}
-          onClose={() => setIsMobileMenuOpen(false)}
-          currentPath={pathname}
-        />
-
-        <div className="flex-1 flex flex-col">
-          {/* PC */}
-          <div className="pc:flex tablet:hidden mobile:hidden h-[50px] border-[#2C2C2C] border-b border-[rgba(255,255,255,0.1)] flex items-center px-5 backdrop-blur-[40px] ">
-            <h1 className="text-[18px] font-bold text-white shadow-[0px_5px_10px_0px_rgba(0,0,0,0.15)]">
-              {getCurrentTitle()}
-            </h1>
-          </div>
-
-          {/* Mobile */}
-          <div className="pc:hidden p-5 border-b border-[rgba(255,255,255,0.1)]">
-            <BackHeader spaceId={spaceId} />
-            <div className="pc:hidden mt-5 h-[30px] flex items-center px-2 hover:bg-[#363636] rounded-lg">
-              <div
-                className="flex justify-between items-center w-full cursor-pointer"
-                onClick={() => setIsMobileMenuOpen(true)}
-              >
-                <h1 className="text-[16px] font-bold text-white shadow-[0px_5px_10px_0px_rgba(0,0,0,0.15)]">
-                  {getCurrentTitle()}
-                </h1>
-                <CaretUpDown size={20} weight="light" className="text-white" />
-              </div>
+        {/* Mobile */}
+        <div className="pc:hidden p-5 border-b border-[rgba(255,255,255,0.1)]">
+          <BackHeader spaceId={spaceId} />
+          <div className="pc:hidden mt-5 h-[30px] flex items-center px-2 hover:bg-[#363636] rounded-lg">
+            <div
+              className="flex justify-between items-center w-full cursor-pointer"
+              onClick={() => setIsMobileMenuOpen(true)}
+            >
+              <h1 className="text-[16px] font-bold text-white shadow-[0px_5px_10px_0px_rgba(0,0,0,0.15)]">
+                {getCurrentTitle()}
+              </h1>
+              <CaretUpDown size={20} weight="light" className="text-white" />
             </div>
           </div>
-
-          <div className="flex-1 overflow-auto p-6">{children}</div>
         </div>
+
+        <div className="flex-1 overflow-auto p-6">{children}</div>
       </div>
-    </>
+    </div>
   );
 };
